flow doctor , admin

+++ Dtos/Staff/AssignDoctorDto.cs
@@ -2,6 +2,6 @@ namespace PetCareSystem.API.Dtos.Staff
 {
     public class AssignDoctorDto
     {
-        public int DoctorId { get; set; }
+        public long DoctorId { get; set; }
     }
 }
